moved class_schedule to the classes folder
